backend/pyproject.toml
[project]
name = "nurtured-heart-api"
version = "1.1.0"
description = "Generate heartfelt compliments using AI and the Nurtured Heart Approach"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "python-dotenv==1.0.1",
    "ollama==0.4.4",
    "python-multipart==0.0.20",
    "email-validator==2.2.0",
    "aiosmtplib==3.0.2",
    "jinja2==3.1.5",
    "google-generativeai==0.8.3",
]

[tool.ruff]
# Set the line length
line-length = 100

# Set the indent width to 2 spaces
indent-width = 2

# Exclude common directories
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "env",
    ".env",
    "*.egg-info",
    "dist",
    "build",
]

# Target Python 3.10+
target-version = "py310"

[tool.ruff.lint]
# Enable standard linting rules
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",   # Line too long (handled by formatter)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use 2 spaces for indentation
indent-style = "space"
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint.isort]
# Sort imports
known-first-party = []

